Util package is a smell, giving more purposed packages started with shell utilities
