+++ src/cmd/setup.rs
@@ -12,8 +12,8 @@ use anyhow::Error;
 use anyhow::Result;
 use argh::FromArgs;
 use lium::util::gen_path_in_lium_dir;
-use lium::util::get_stdout;
-use lium::util::run_bash_command;
+use lium::util::shell_helpers::get_stdout;
+use lium::util::shell_helpers::run_bash_command;
 
 #[derive(FromArgs, PartialEq, Debug)]
 /// setup development environment
